+++ cwLee/README.md
@@ -2,7 +2,7 @@
 
 ---
 
-## 1주차
+## 10주차
 
 ### 5.7
 
@@ -47,5 +47,14 @@
 2. font.py를 실행시키면 pygame의 폰트를 확인할 수 있습니다.
 3. 캐릭터가 특정 좌표 못넘어가도록 벽이랑 충돌대신 구현했습니다.
 4. (공부 아이콘)문 앞에서 space키를 누르면 다음 스테이지로 넘어가도록 구현했습니다.
+
+---
+
+##11주차
+
+### 5.15
+
+1. 선택분기에 따른 텍스트 출력파일을 만들어 보았습니다.
+   - q, e를 눌러 분기를 선택하면 분기에 따른 텍스트가 출력됩니다.
 
 ---
