+++ font.py
@@ -1,0 +1,3 @@
+import pygame
+
+print(pygame.font.get_fonts())
