+++ img/Room_test/Room1_floor.png
