--- story.md
@@ -1,66 +1,0 @@
-# <타이틀 화면(제목)>
-
----
-
-## <프롤로그>
-
-    어서오세요./
-    게임에 참가하게 된 것을 환영합니다./
-    방법은 지극히 간단합니다. 어렵지 않아요./
-    앞으로 마주할 커다란 갈림길 앞에서…/
-    당신은 하나의 선택을 하게 될 것입니다./
-    그리고 그 선택이 서로 다른 이야기를 만들고./
-    서로 다른 결말을 낳겠죠./
-    ……우리 모두가 그렇듯이./
-    그럼, 출발합시다./
-
----
-
-## <소년기>
-
-    당신은 무럭무럭 자라 다른 아이들과 마찬가지로 학교에 다니게 되었습니다. 당신의 관심사는?/
-    (공부 아이콘 선택)>무언가를 배우고 좋은 성적을 받는 것을 좋아하는군요./
-    (예체능 아이콘 선택)>당신의 끼를 발산하는 것을 좋아하는군요./
-
----
-
-## <청년기>
-
-    스스로의 관심사에 맞추어 대학에 진학한 당신. 청춘의 한가운데 당신이 열중하는 것은?/
-    (전공 아이콘 선택)>당신은 꿈을 좇아 멈추지 않는군요./
-    (연애 아이콘 선택)>청춘은 불같은 사랑 없이는 성립하지 않죠./
-
----
-
-## <장년기>
-
-    인생의 반환점을 지나온 당신. 이제는 어디에 충실해야 할지 알 것 같은데?
-    (일 아이콘 선택)>성공으로 보답하는 것만이 최고의 선택이죠.
-    (가정 아이콘 선택)>역시 곁에 있어 주는 가족이 최고죠.
-
----
-
-## <노년기(결말)>
-
-    수고하셨습니다./
-    ……슬슬 엔딩이 다가오고 있어요./
-    당신은……./
-
-    (공부+전공+일 선택)>어릴 때부터 확고한 관심사를 가졌던 당신은 자연스레 목표를 이루기 위해 최선을 다했어요. 그 결과 최고의 자리까지 올라갔지만……조금은 쓸쓸한 인생이었을지도 모르겠네요. 수고했어요.
-    (공부+전공+가정 선택)>어릴 때부터 확고한 관심사를 가졌던 당신은 청춘을 불사르며 열심히 노력했어요. 하지만 가족의 소중함 역시 깨달은 것 같네요. 나쁘지 않아요. 수고했어요.
-    (공부+사랑+일 선택)>학구열이 강한 당신은 열심히 공부하여 대학에 진학했고, 그곳에서 진정한 사랑을 만나 뜨거운 청춘을 보냈군요. 나이가 들어선 다시 일에 열중했지만, 그래도 당신은 젊은 날을 추억하며 미소짓고 있네요.
-    (공부+사랑+가정 선택)>학구열이 강한 당신은 열심히 공부하여 대학에 진학했지만, 진정한 사랑을 만나 뜨거운 청춘을 보냈군요. 나이가 들어서도 가정에 충실하며 좋은 부모이자 배우자였네요. 당신의 삶은 분명 꽤 따뜻했겠네요.
-    (예체능+전공+일 선택)>통통 튀는 재능이 있는 당신은 청춘을 모두 바쳐 세계적으로 유명한 인재가 되었군요. 지금은 당신의 삶을 모두가 우러러보고 있어요. 훌륭해요.
-    (예체능+전공+가정 선택)>통통 튀는 재능이 있는 당신은 청춘을 바쳐 그 재능에 충실했지만, 나이가 들어선 그 이상으로 가치있는 것을 찾아냈군요. 가족의 품 만한 게 또 없죠.
-    (예체능+사랑+일 선택)>주체할 수 없는 끼를 가진 당신은 그 열정을 연인에게 쏟기로 결심했어요. 나이가 들어선 다시 일에 충실했지만, 그 청춘의 한 페이지가 어디 가는 것은 아니죠.
-    (예체능+사랑+가정 선택)>주체할 수 없는 끼를 가진 당신은 그 열정을 연인에게 쏟기로 결심했고, 나이가 들어서도 사랑하는 사람에겐 늘 최선을 다했군요. 당신의 선택을 존중하고 축복해요.
-
-
-    이상입니다./
-    즐거운 삶이었나요?/
-    부디 그러셨길 바랄게요. 정답은 없으니까요./
-    그래도 다시 한번 돌아간다면……../
-
-    >다시 도전한다/게임을 끝낸다
-
----
