+++ story.txt
